fix ffi-check on GIL-enabled build

+++ pyo3-ffi/src/object.rs
@@ -36,9 +36,9 @@ pub struct PyObjectObFlagsAndRefcnt {
 /// This struct is anonymous in CPython, so the name was given by PyO3 because
 /// Rust structs need a name.
 pub struct PyObjectObFlagsAndRefcnt {
-    pub ob_refcnt: u16,
+    pub ob_refcnt: u32,
     pub ob_overflow: u16,
-    pub ob_flags: u32,
+    pub ob_flags: u16,
 }
 
 #[repr(C)]
